Remove binary and function cache, use Ghidra property storage instead

+++ src/main/java/ai/reveng/toolkit/ghidra/core/services/api/types/FunctionID.java
@@ -1,4 +1,4 @@
 package ai.reveng.toolkit.ghidra.core.services.api.types;
 
-public record FunctionID(Integer value) {
+public record FunctionID(long value) {
 }
